server/requirements.txt
flask==3.0.0
flask-cors
hanzidentifier
jieba
openai-whisper
xpinyin
